refactor: Campanha SDR renderiza direto na aba (sem etapa extra)
Extrai a tela de config para o componente CampaignConfig e renderiza-o direto na aba
"Campanha SDR" de Configurações (some o card "Abrir configuração" + o hop). Remove o
breadcrumb. /sdr-ia/parametros vira redirect p/ /settings?tab=campanha-sdr; lib/modules.ts
aponta o módulo pro novo caminho. Função da tela 100% preservada.
